.agent/skills/summarize/scripts/requirements.txt
google-genai
python-dotenv
youtube-transcript-api==1.2.3
cloudscraper
beautifulsoup4
